Enable NarrowCalculation error-prone check
This mostly matches IntelliJ's `integer multiplication implicitly cast
to long` inspection. While places changed do not look like bugs, so can
be viewed as false positives, this helps avoid IDE warnings and thus
makes spotting problems easier.

+++ lib/trino-orc/src/main/java/io/trino/orc/stream/LongInputStreamV1.java
@@ -88,7 +88,7 @@ public class LongInputStreamV1
             readValues();
         }
         if (repeat) {
-            result = literals[0] + (used++) * delta;
+            result = literals[0] + (used++) * (long) delta;
         }
         else {
             result = literals[used++];
@@ -111,7 +111,7 @@ public class LongInputStreamV1
             int chunkSize = min(numLiterals - used, items);
             if (repeat) {
                 for (int i = 0; i < chunkSize; i++) {
-                    values[offset + i] = literals[0] + ((used + i) * delta);
+                    values[offset + i] = literals[0] + ((used + i) * (long) delta);
                 }
             }
             else {
@@ -138,7 +138,7 @@ public class LongInputStreamV1
             int chunkSize = min(numLiterals - used, items);
             if (repeat) {
                 for (int i = 0; i < chunkSize; i++) {
-                    long literal = literals[0] + ((used + i) * delta);
+                    long literal = literals[0] + ((used + i) * (long) delta);
                     int value = (int) literal;
                     if (literal != value) {
                         throw new OrcCorruptionException(input.getOrcDataSourceId(), "Decoded value out of range for a 32bit number");
@@ -177,7 +177,7 @@ public class LongInputStreamV1
             int chunkSize = min(numLiterals - used, items);
             if (repeat) {
                 for (int i = 0; i < chunkSize; i++) {
-                    long literal = literals[0] + ((used + i) * delta);
+                    long literal = literals[0] + ((used + i) * (long) delta);
                     short value = (short) literal;
                     if (literal != value) {
                         throw new OrcCorruptionException(input.getOrcDataSourceId(), "Decoded value out of range for a 16bit number");
